requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
geographiclib==1.50
geopy==1.20.0
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.1
pandas==0.25.3
pkg-resources==0.0.0
pyspellchecker==0.5.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.0
WTForms==2.2.1